moved to sql.function.string renamed to CastTest edited 1 line in SqlTest
